combine modules
